config files refactor + features
This brings some fairly major changes relating to how server configuration is
managed.

Refactored:

- now uses a json file to store settings
- separate module for managing configuration instead of just importing a python
  config file
- config module keeps the same api, so all current server code accessing config
  values isn't modified

New Features:

- all config has a default value configured, so if the user wishes to have a
  configuration file with the minimum json `{}` in it, it will all still work.
- able to pass a config file in the command line args to specify alternate
  config files - great for quick testing/debugging (plus easy to add new command
  line args if we want)
