+++ config.json
@@ -1,0 +1,16 @@
+{
+        "database_url": "sqlite:///db.sqlite3",
+        "cors_origin": "*",
+        "listen_port": 8888,
+        "debug_mode": true,
+        "serve_webapp": true,
+        "on_rpi": false,
+        "test_data": null,
+        "sensor_params": null,
+        "temp_interval": 300,
+        "temp_max_length": 300,
+        "server_name": "the temperature server",
+        "location": null,
+        "timezone": null
+}
+
